restructure and extra components
